# docs/swagger/swagger.yaml
basePath: /
definitions:
  myaaw_internal_common.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  myaaw_internal_pkg.Chat:
    properties:
      first_name:
        type: string
      id:
        type: integer
      type:
        type: string
      username:
        type: string
    type: object
  myaaw_internal_pkg.Document:
    properties:
      file_id:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_unique_id:
        type: string
      mime_type:
        type: string
      thumb:
        $ref: '#/definitions/myaaw_internal_pkg.Thumbnail'
      thumbnail:
        $ref: '#/definitions/myaaw_internal_pkg.Thumbnail'
    type: object
  myaaw_internal_pkg.From:
    properties:
      first_name:
        type: string
      id:
        type: integer
      is_bot:
        type: boolean
      language_code:
        type: string
      username:
        type: string
    type: object
  myaaw_internal_pkg.Photo:
    properties:
      file_id:
        type: string
      file_size:
        type: integer
      file_unique_id:
        type: string
      height:
        type: integer
      width:
        type: integer
    type: object
  myaaw_internal_pkg.ReplyToMessage:
    properties:
      caption:
        type: string
      chat:
        $ref: '#/definitions/myaaw_internal_pkg.Chat'
      date:
        type: integer
      document:
        $ref: '#/definitions/myaaw_internal_pkg.Document'
      from:
        $ref: '#/definitions/myaaw_internal_pkg.From'
      message_id:
        type: integer
      photo:
        items:
          $ref: '#/definitions/myaaw_internal_pkg.Photo'
        type: array
      text:
        type: string
      voice:
        $ref: '#/definitions/myaaw_internal_pkg.Voice'
    type: object
  myaaw_internal_pkg.TelegramIncommingChat:
    properties:
      message:
        $ref: '#/definitions/myaaw_internal_pkg.UserMessage'
      update_id:
        type: integer
    type: object
  myaaw_internal_pkg.TelegramSendMessageStatus:
    properties:
      description:
        type: string
      error_code:
        type: integer
      ok:
        type: boolean
      result:
        $ref: '#/definitions/myaaw_internal_pkg.UserMessage'
    type: object
  myaaw_internal_pkg.Thumbnail:
    properties:
      file_id:
        type: string
      file_size:
        type: integer
      file_unique_id:
        type: string
      height:
        type: integer
      width:
        type: integer
    type: object
  myaaw_internal_pkg.UserMessage:
    properties:
      caption:
        type: string
      chat:
        $ref: '#/definitions/myaaw_internal_pkg.Chat'
      date:
        type: integer
      document:
        $ref: '#/definitions/myaaw_internal_pkg.Document'
      from:
        $ref: '#/definitions/myaaw_internal_pkg.From'
      message_id:
        type: integer
      photo:
        items:
          $ref: '#/definitions/myaaw_internal_pkg.Photo'
        type: array
      reply_to_message:
        $ref: '#/definitions/myaaw_internal_pkg.ReplyToMessage'
      text:
        type: string
      voice:
        $ref: '#/definitions/myaaw_internal_pkg.Voice'
    type: object
  myaaw_internal_pkg.Voice:
    properties:
      duration:
        type: integer
      file_id:
        type: string
      file_size:
        type: integer
      file_unique_id:
        type: string
      mime_type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Myaaw - Integrate your favorite LLM with a Telegram bot.
  title: Myaaw API
  version: "1.0"
paths:
  /webhook/bot:
    post:
      consumes:
      - application/json
      description: To receive incoming message from RabbitMQ consumer
      parameters:
      - description: Telegram incoming chat
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/myaaw_internal_pkg.TelegramIncommingChat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/myaaw_internal_pkg.TelegramSendMessageStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/myaaw_internal_common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/myaaw_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/myaaw_internal_common.ErrorResponse'
      summary: Bot
      tags:
      - Bot
  /webhook/telegram:
    post:
      consumes:
      - application/json
      description: To receive incoming message from Telegram and push to Queue
      parameters:
      - description: Telegram incoming chat
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/myaaw_internal_pkg.TelegramIncommingChat'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/myaaw_internal_common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/myaaw_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/myaaw_internal_common.ErrorResponse'
      summary: Queue
      tags:
      - Bot
swagger: "2.0"